Create Reverse
Websites usually ask you to input strings in Hebrew or in any other languages, But once you do that, they invert your text for example:

What a wonderful day --> what you wrote
day wonderful a What --> what the website displays

Use this function to invert the text you what to write, and it will display a text that is backwards to it.

So now you will be able to copy and paste the already "inverted" text, into the website, essentially writing it normally.

For optimization reasons, You can input a text file into the function. Works in Hebrew or English.